tracker with small group of tests

+++ src/main/java/ru/spbau/mit/torrent/io/query/UpdateQuery.java
@@ -3,10 +3,10 @@ package ru.spbau.mit.torrent.io.query;
 import java.util.List;
 
 public class UpdateQuery extends Query {
-    public final short port;
+    public final int port;
     public final List<Integer> fileIds;
 
-    public UpdateQuery(short port, List<Integer> fileIds) {
+    public UpdateQuery(int port, List<Integer> fileIds) {
         super(QueryId.UPDATE);
         this.port = port;
         this.fileIds = fileIds;
